feat: article 및 comment service 코드 일부 기능 함수화
