+++ express-mission-prisma/src/services/articleService.js
@@ -1,60 +1,21 @@
 import articleRepository from "../repositories/articleRepository.js";
+import {
+  createFillterOtionsByKeyword,
+  createPageSizeFilterOptions,
+} from "../utils/filterOptions.js";
 
 async function create(createData) {
   return await articleRepository.create(createData);
 }
 
 async function getAllByFilter(query) {
-  const { page, pageSize, orderBy, keyWord = "" } = query;
-
-  const pageNum = page || 1;
-  const pageSizeNum = pageSize || 10;
-  const order = orderBy || "recent";
-  const offset = (pageNum - 1) * pageSizeNum;
-  const whereOrBody = {
-    contains: keyWord,
-    mode: "insensitive",
-  };
-  const whereOr = {
-    OR: [
-      {
-        title: whereOrBody,
-      },
-      {
-        content: whereOrBody,
-      },
-    ],
-  };
-
-  const fillter = {
-    orderBy: { createdAt: "desc" },
-    skip: parseInt(offset),
-    take: parseInt(pageSizeNum),
-    where: whereOr,
-  };
-
-  return await articleRepository.getAllByFilter(fillter);
+  const filterByPageSize = createPageSizeFilterOptions(query);
+  return await articleRepository.getAllByFilter(filterByPageSize);
 }
 
 async function countByFilter(query) {
-  const { keyWord = "" } = query;
-
-  const fillterBody = {
-    contains: keyWord,
-    mode: "insensitive",
-  };
-  const fillter = {
-    OR: [
-      {
-        title: fillterBody,
-      },
-      {
-        content: fillterBody,
-      },
-    ],
-  };
-
-  return await articleRepository.countByFilter(fillter);
+  const fillterByKeword = createFillterOtionsByKeyword(query);
+  return await articleRepository.countByFilter(fillterByKeword);
 }
 
 async function getById(id) {
@@ -63,7 +24,6 @@ async function getById(id) {
 
 async function update(id, updateData) {
   const updateDataWithId = { where: { id }, data: updateData };
-
   return await articleRepository.update(updateDataWithId);
 }
